Added supported bitfields to the device.
Added the bitfield items to the device, only test the bitfield
constructor, still need to add more. Added the option of converting the
data into the byte array, to be ready to send. still in progress.  Each
enum in the end will need to be able to convert data into the byte
array, and vise versa.