Added get Device Info command and status
I added the command and status for the getDeviceInfo command(0x81) this
will fetch the information like sw and hw version number, and all the
supporting bitfields. It runs through a bunch of checks to make sure it
works.
Still need to add a lot of tests to validate this.